api_project/requirements.txt
Django==5.2
djangorestframework==3.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
django-cors-headers==4.3.1 